Refactor functions into DucklingRecognizer
